backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
python-dotenv>=1.0,<2.0